move panel creation to utils
