+++ src/rtasr/datasets/dataset_ami.py
@@ -20,8 +20,8 @@ from rich import print
 from rich.live import Live
 from rich.progress import Progress, TaskID
 
-from rtasr.constants import DATASETS, create_live_panel
-from rtasr.utils import download_file, get_files, resolve_cache_dir
+from rtasr.constants import DATASETS
+from rtasr.utils import create_live_panel, download_file, get_files, resolve_cache_dir
 
 
 async def prepare_ami_dataset(output_dir: str = None, use_cache: bool = True) -> None:
